% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_points_usa.R
\name{grid_points_usa}
\alias{grid_points_usa}
\title{Create a grid of points that fall within the contiguous USA}
\usage{
grid_points_usa(resolution = 100)
}
\arguments{
\item{resolution}{numeric value for resolution of map}
}
\value{
A dataframe of grid points with longitude, latitude, and whether it's inside the USA
}
\description{
Get a data frame of grid points. Each row corresponding to a specific grid point,
showing its longitude, latitude, and whether it's inside the USA by binary.
}
\examples{
#grid with resolution 30
grid_points_usa(30)

#grid with resolution 100
grid_points_usa()
}
