% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stations}
\alias{stations}
\title{Latitude/Longitude Coordinates of Weather Stations in the USA}
\format{
a dataframe with
\describe{
\item{WBANNO}{The station WBAN number}
\item{state}{The station's located state }
\item{station_name}{Station name}
\item{LONGITUDE}{Longitude of station in decimal degrees }
\item{LATITUDE}{Latitude of station in decimal degrees}
}
}
\usage{
stations
}
\description{
A dataset with the lognitude latitude coordinates of weather stations in USA
}
\keyword{datasets}
