% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature_trend.R
\name{temp_trend}
\alias{temp_trend}
\title{Get the overall temperature trend in the US contiguous 48}
\usage{
temp_trend(station_id = NULL)
}
\arguments{
\item{station_id}{optional numeric WBANNO number for specified station}
}
\value{
a linear model of class merMod or lm
}
\description{
Get a model of temperature change over time in degrees Celsius per year
for the contiguous 48 states, based on USCRN data.
}
\examples{
# all stations
tt_all <- temp_trend()
summary(tt_all)

# for Ithaca
tt_ithaca <- temp_trend(64758)
summary(tt_ithaca)
}
