% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station_analysis.R
\name{time_series}
\alias{time_series}
\title{Create a time series for a USCRN weather station}
\usage{
time_series(station_id, start_date = NULL, end_date = NULL)
}
\arguments{
\item{station_id}{numeric station ID (WBANNO number)}

\item{start_date}{character string with format YYYY-mm-dd for start of time series}

\item{end_date}{character string with format YYYY-mm-dd for end of time series}
}
\value{
A data frame containing the full time series from the start of data collection to the most recent date, or between user-specified dates.
}
\description{
Get every entry recorded by USCRN for a given weather station, between
given start and end dates (default: the entire time series) in a data frame.
}
\examples{
ithaca <- 64758
ithaca_ts <- time_series(ithaca)
ithaca_2010s <- time_series(ithaca, start_date = "2010-01-01",
                            end_date = "2019-12-31")
}
