% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearly_temp_specific.R
\name{yearly_cycle}
\alias{yearly_cycle}
\title{Model the yearly temperature cycle of a specified USCRN weather station}
\usage{
yearly_cycle(station_id)
}
\arguments{
\item{station_id}{The station ID (WBANNO number)}
}
\value{
A data frame containing two columns:
\itemize{
\item \code{doy} the day of year (1-365)
\item \code{t_daily_pred} predicted average temperatures for each day, in degrees Celsius
}
}
\description{
Get predicted values for each day of the year based on a sinusoidal model
fit on the USCRN average daily temperatures for a given research station.
}
\examples{
ithaca <- 64758
ithaca_cycle <- yearly_cycle(ithaca)
plot(t_daily_pred ~ doy, data = ithaca_cycle, type = "l")
}
