% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_points_usa.R
\name{grid_points_usa}
\alias{grid_points_usa}
\title{Create a grid of points that fall within the contiguous USA}
\usage{
grid_points_usa(resolution = 100)
}
\arguments{
\item{resolution}{integer value specifying the number of grid divisions}
}
\value{
A dataframe of grid points with longitude, latitude, and whether the point is inside the USA
\itemize{
\item \code{long} Longitude in decimal degrees
\item \code{lat} Latitude in decimal degrees
\item \code{in_usa} numeric 1 if the point is in the contiguous USA, 0 otherwise
}
}
\description{
Get a data frame of grid points. Each row corresponds to a specific grid point,
containing its longitude, latitude, and whether it's inside the contiguous
United States.
}
\details{
The longitudes and latitudes correspond to an \eqn{n \times n} grid with
points spanning the contiguous USA, with \eqn{n =} \code{resolution}
}
\examples{
#grid with resolution 30, 900 points in total
grid_points_usa(30)

#grid with default resolution (100), 10,000 points in total
grid_points_usa()
}
