% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolation.R
\name{interpolation}
\alias{interpolation}
\title{Interpolate data from the stations to a grid points within the contiguous USA}
\usage{
interpolation(value, resolution)
}
\arguments{
\item{value}{a vector of data corresponding to 236 station locations}

\item{resolution}{numeric value for resolution of map}
}
\value{
A data frame of prediction with its corresponding longitude and latitude
}
\description{
Interpolate data from the stations to a grid points within the contiguous USA
}
\examples{
#To get data frame of predicted average year temperature of USA with resolution 100
interpolation( Average temperature of each station, 100)

#To get data frame of predicted temperature in March 2024 of USA at each long, lat with resolution 30
interpolation(March 2024 average temperature of each station, 30)
}
