% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_interpolation.R
\name{plotting_interpolation}
\alias{plotting_interpolation}
\title{Plot the interpolated data}
\usage{
plotting_interpolation(value, stations = NULL, resolution = 100)
}
\arguments{
\item{value}{a vector of data corresponding to stations specified in \code{stations}}

\item{stations}{an optional numeric vector of station WBANNOs to use in interpolation}

\item{resolution}{numeric value for grid resolution}
}
\value{
\code{NULL}
}
\description{
Using the interpolated data, produces a visual representation of predicted
results by plotting through image.plot
}
\examples{
# visualize predicted estimated trend of temperature of USA with
# resolution 100
tt_all <- temp_trend(rustcrn::stations$WBANNO)
value  <- tt_all[, 2]
plotting_interpolation( value )

# visualize interpolation of every other station at a lower resolution
other <- rustcrn::stations$WBANNO[seq(2, nrow(rustcrn::stations), by = 2)]
value_other <- temp_trend(other)[,2]
plotting_interpolation( value_other, stations = other, resolution = 75 )
}
