% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_interpolation.R
\name{plotting_interpolation}
\alias{plotting_interpolation}
\title{Plot the interpolated data}
\usage{
plotting_interpolation(value, resolution)
}
\arguments{
\item{value}{a vector of data corresponding to 236 station locations}

\item{resolution}{numeric value for resolution of map}
}
\value{
Produce visual representation of spatial predictions
}
\description{
Using the interpolated data, produces a visual representation of predicted
results by plotting through image.plot
}
\examples{
#For image of predicted average year temperature of USA with resolution 100
plotting_interpolation(value, resolution)

}
