% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_interpolation.R
\name{plotting_interpolation}
\alias{plotting_interpolation}
\title{Plot interpolated USCRN data on the US contiguous 48}
\usage{
plotting_interpolation(
  interpolation,
  main = "Predicted Value by Location in US",
  palette = "Terrain",
  lines = TRUE
)
}
\arguments{
\item{interpolation}{the output of [rustcrn::interpolation()]; a data frame containing four columns \code{long, lat, in_usa, prediction}}

\item{main}{plot title, default is "Predicted Value by Location in US"}

\item{palette}{a valid palette name to be passed to [grDevices::hcl.colors()]}

\item{lines}{a logical indicating whether a map of the US should be drawn on top of the map}
}
\value{
\code{NULL}
}
\description{
Using interpolated weather station data, draw a visual representation of the
predicted results in the form of a heat map-style plot of the US contiguous 48
states.
}
\details{
See [rustcrn::interpolation()], [fields::imagePlot()], [grDevices::hcl.colors()].
}
\examples{
# visualize a dummy dataset interpolated across the US
sub <- rustcrn::stations[ seq(1, nrow(rustcrn::stations), by = 4), ]
st <- as.numeric(as.factor(sub$state))
st_in <- interpolation(st ~ LONGITUDE + LATITUDE, data = sub,
                       resolution = 20)
plotting_interpolation( st_in, main = "Interpolation by state",
                        palette = "Spectral", lines = TRUE )
}
