% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature_trend.R
\name{temp_trend}
\alias{temp_trend}
\title{Get temperature trends for stations in the US contiguous 48}
\usage{
temp_trend(station_ids)
}
\arguments{
\item{station_ids}{vector of WBANNO numbers to get trends for specified stations.}
}
\value{
a data frame with four columns and n rows, where n is the number of stations in the input vector
\itemize{
\item \code{WBANNO} the station ID
\item \code{trend} trend of temperature change, in degrees Celsius per year
\item \code{p} the p-value of the year coefficient
\item \code{se} standard error of the model
\item \code{len} the length of data collection, in years
}
}
\description{
Get a temperature change over time in degrees Celsius per year
for the contiguous 48 states, based on USCRN data.
}
\details{
Temperature trends are modeled using a linear model with year and sinusoidal
components, to account for seasonality, and the function returns a small
summary of the model results including the coefficient for the year component
(the trend).
}
\examples{
# all stations
tt_all <- temp_trend(rustcrn::stations$WBANNO)

# for Ithaca
tt_ithaca <- temp_trend(64758)
tt_ithaca[["trend"]]
}
