% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station_analysis.R
\name{time_series}
\alias{time_series}
\title{Create a time series for a USCRN weather station}
\usage{
time_series(station_id, start_date = NULL, end_date = NULL)
}
\arguments{
\item{station_id}{The station ID (WBANNO number)}

\item{start_date}{Start of time series (format YYYY-mm-dd)}

\item{end_date}{End of time series (format YYYY-mm-dd)}
}
\value{
A data frame containing the full time series from the start of data collection to the most recent date, or between user-specified dates.
}
\description{
Get every entry recorded by USCRN for a given weather station, between
given start and end dates (default: the entire time series) in a data frame.
}
\examples{
ithaca <- 64758
ithaca_cycle <- yearly_cycle(ithaca)
plot(t_daily_pred ~ doy, data = ithaca_cycle, type = "l")
}
